[twisty] Add a hack to avoid oversized `Twisty3DCanvas`es.
This *hopefully* fixes #73.

Note that `position: relative` for the wrapper and `position: absolute` is the only combination that works. However, it allows us to tell `THREE.js` to set both the canvas pixel size and the CSS display size (whose ratio is the pixel density of the screen) in `setSize()`.