[twisty] Disable orbit control zoom for now.
We should make sure to size the puzzles properly so that zooming is not necessary.

We may want to support zoom in the long term, but we may want to reserve scrolling for something else (alg.cubing.net currently uses it to change view angle)..

Client code can reach into the `threeOrbitControls` to enable zoom for now, if needed.