[twisty] Cursor: fold completed moves into state.
This avoids the old behaviour where the cursor would say there was a
move in progress at at the end of the timeline, and never dispatched the
end state. For some applications, it's nice to have the actual end state
without having to do extra calculations.

It would be nice to take this further and actually dispatch the
"closest" state for a move in progress. But this would take careful work
to accommodate non-quantum moves, and doesn't necessarily make sense
when there are simultaneous moves. An eventual workaround might be to
give listeneres a way to query the closest state for the dispatched
timestamp.