[puzzles] Move KPuzzle definitions into a new package.
This package uses `Promise`s so that an implementation for any given puzzle can be loaded when it is needed, instead of requiring all puzzles to be loaded when the library is initially loaded. This will also help to avoid loading `THREE.js` if it is not needed for a given page.