[puzzle-geometry + puzzles] Add Master Tetraminx.
This also restores vertex-only-turning tetrahedra to pure SiGN notation,
with the sole exception of a WCA hack to swap e.g. R ↔ r. It's unclear what
notation is best to use for e.g. Master Pyraminx, so we punt on this by
adding Master Tetraminx to `cubing/puzzles` first.

However, we still want to keep the same family mapping as Pyraminx, so
we introduce a class adjacent to `PyraminxNotationMapper` that invokes
the (simpler) tetraminx mapping.