[twisty] Fix a thundering herd bug in `TwistyPropDerived`.
Previously, we were conflating "last successful calculation that has
finished" with "latest calculation that has started". This commit splits
those into two separately cached fields.

This is super important for prop update that has lots of paths to
listeners, since all those listeners will be notified and ask for the
prop almost immediately. We weren't deduplicating those requests before,
but now we properly use the generation number to do so, while also
allowing the last calculation to be reused if the inputs haven't
changed.

We could certainly add a few more clever tricks (e.g. caching the
second-to-last successful result, in case something is rapidly toggled),
but this commit should already be a massive improvement for some
situations.