[tsconfig.json] Attempt to exclude `dist`.
Hopefully closes #133 but probably not.

VSCode seems to have this infuriating insistence on autocompleting from
the `dist` dir and ignoring the `exclude` field of `tsconfig.json`. In
fact, VSCode reliably *prefers* these completions to the actual sources in
`src/cubing` and uses them by default.

I've wasted a lot of hours on this and still don't have reliable
results, but this change applies aggressive scoping to our config to
make it really clear we don't want to include `dist`. This may
accidentally exclude some files outside `src` by accident, but hopefully
we can deal with that later.