[package.json] Exclude the Chromium download from Puppeteer in `make initial-setup`.
This still leaves us with a ridiculous number of files in `node_modules`
that you don't need for `make dev`, but it cuts down on a lot of the
total file size.

Ideally we'd install just the ones we need for `make dev`. but `npm`
doesn't provide a simple portable way to do this, and in particular has
decided against supporting optional dev dependencies in the foreseeable
future: https://github.com/npm/npm/issues/3870