[twisty] Bump the default dedicated renderer limit to 2.
This allows a player with front + back view to grab a dedicated render for each (which is important for Twizzle Explorer). I'm not aware of any other way to get acceptable performance in Safari otherwise.

If there are no back views, this allows the first two players to nab a dedicated renderer each (but not their back views, if those are enabled later). That's an acceptable tradeoff for the simplicity of this heuristic, which *hopefully* won't be necessary in a few months/years when Safari provides a reasonable way to get performance with shared renderers.