[search] Use a module worker where possible.
Unfortunately, there is no robust way to do this until `esbuild` supports web workers: https://github.com/evanw/esbuild/issues/312

This is because the instantiation site needs to know the URL of the worker entry file, which is an unpredictable hash.

This commit introduces a workaround: ask the entry file for its URL. That is, we add a wrapper entry file that *either* starts the worker code *or* reports its URL, depending on what it's asked. Assuming that bundlers preserve the semantics of the entry file hierarchy for dynamic imports, this should be broadly compatible in module environments.