[twisty] Move *all* `three` imports behind a dynamic import.
This removes `three` from blocking the load of
http://localhost:3333/alpha.twizzle.net/edit/?puzzle=clock , taking it
from ≈2MB to ≈800KB.

For 3D puzzles, It also allows the player UI to show while the 3D code
is still loading. That allows going full screen, or even pressing play —
all before the puzzle is visible. It opens up the possibility of
fallbacks, or for an app to switch to a 2D puzzle without waiting for
the 3D puzzle to load first.

Unfortunately, the load of the bundle containing `three` is delayed
until it is requested — in theory we could use even cleverer splitting
to detect as soon as possible (e.g. based in the `visualization`
parameter of a player) if that bundle will be needed, and begin the
download in parallel.

Also note that there are no mechanisms or tests to enforce this yet. I
want to work on that at some point, but it's great that we've already
made it possible.