[search] Introduce cubing-private exports for `search`.
This allows us to use exports across sub-packages without exporting them publicly. This is generally a sign of bad design, but it allows us to ship certain functionality without exposing an experimental API that we do not intend to support in the future.

In particular, this commit removes `[experimental]randomScrambleForEvent()` from `cubing/search`, since it's only meant to be used from `cubing/scramble`.