[puzzles] Move `experimentalStickerings` into `cubing/puzzle`.
This is a bit awkward, since similar configs still remain in
`cubing/twisty`. But we need to move it out to avoid a cyclic import,
and eventually all puzzle-specific info (such as stickering definitions)
should be in `cubing/puzzles` anyways. So this particular move is really
just an early start on that migration.