TypeScript likes to suggest things like: `Add import from "cubing/dist/types/alg"` when `node_modules` are included in the project.
When using `"module": "nodenext"` in `tsconfig.json`, TypeScript uses `cubing/alg` as intended.

This commit was an attempt to fix this for the TypeScript compiler
defaults, by actually placing files in the legacy folders that are
published in the root package. That did not work at all.

I think TypeScript actually looks for the exported names themselves,
which are only present in the `dist` folder. I don't want to move any
`dist` files into the legacy folders, so I'm out of ideas. I might just
have to wait until TypeScript's default behaviour matches `nodenext` for
this.