Revert "[puzzles] Export the exact `cube3x3x3` type."
This reverts commit 38eba5d3b13d93568adeb48f3dfac1ed3cebcff6.

`npx web-test-runner` seems to be unable to recognize the `satisfies`
operator, and I can't figure out what part of it lacks support. This is
a bogus excuse for a revert, but I don't have time to figure out the
details and I want to get tests passing on `main` again.