[experiments/svg-state-editor] Clone the start state before editing.
Previously, we were editing the original start state, which means that the definition was modified as well. This meant that if you downloaded a def and a scramble at the same time, `twsearch` would always find that it's already solved. 😆

Ideally we'd find a way to enforce read-only access to prevent modifying existing states by default, but that could be expensive.