[search] Move `search-worker-entry.js` back into the module.
We can use an object with `in`/`out` for `esbuild` to place `search-worker-entry.js` in the right place for the `esm` build instead of placing the file at the project root.

This is not ideal, but it allows the source code to use correct relative paths instead of relying on `esbuild` implicitly. If `esbuild` supports `import.meta.resolve(…)` or `new URL(…, import.meta.url)` at some point, this should also continue to work without additional changes (although we'd be able to remove the search worker entry from `entryPoints` for the `esm` and `bin` builds).