[types] Include re-exported types in top-level folders again.
This allows TypeScript to pick up the types without having to author `tsconfig.json`, particularly in VS Code.

I want to get rid of this as soon as possible, as using anything other than the `"exports" field in `package.json` is an anti-pattern. But this is going to provide a much better experience in the meantime.