[test] Add a regression test for af5fa032 by testing for `globalThis.__THREE__`.
In theory `esbuild` could split `three.js` across
bundles, but in practice it seems to be mostly in
a single file. So `globalThis.__THREE__` is a good
signal for now.

In the long term, we should actually test the
number of bytes sent over the wire for the page
load.