[Makefile] Explicitly install `playwright`.
(Is this why we were using Puppeteer? Why can't `@web/test-runner` install Playwright automatically?)

Note that this is *much* slower than `make
update-dependencies` in the no-op case. Browser
tests are already a bit slow, so we can kind of
tolerate this by only running it as a dependency
for browser tests. But it's also disappointing
that this is so slow, and also means that `make
update-dependencies` doesn't leave `make test-all` in a state
that can be run offline. 😕