[bluetooth] Add optional MAC retrieval function
To work with many Bluetooth smart cubes, it is required to know the
cube's MAC address, which is not accessible through Web Bluetooth. For
instance, each of the GAN Gen2 smart cubes use the cube's MAC address to
salt the encryption key and IV. While it is possible to get this through
`watchAdvertisements()`, this doesn't work on all platforms (Linux and
ChromeOS are not supported).

If applied, this commit adds a parameter that is passed to each cube's
connect function, so that cubes that need this information may choose to
request it. To keep this functionality simple, a request function can be
passed by a user calling the `connectSmartPuzzle()` function to, for
example, prompt the user for the cube's MAC address with a custom UI.
This commit also adds a simple default function which uses the browser's
default `prompt()`.